Added a few isVisible() checks in the logic of two widgets and a container, as well as fixing a potential stack corruption on client exit which was only visible when logic was disabled in the progress bar when not visible.

+++ src/bindings/guichan/widgets/progressbar.h
@@ -131,7 +131,7 @@ class ProgressBar : public gcn::Widget
         bool mSmoothProgress;
 
         std::vector<gcn::Color*> mColors;
-        int mCurrentColor;
+        size_t mCurrentColor;
 
         Uint8 mRed, mGreen, mBlue;
         Uint8 mRedToGo, mGreenToGo, mBlueToGo;
